Add basic error handling
